Retrieve direct dependencies for MSBuild task
Previously, the ReferenceCop task retrieved transitive dependencies
by initiating a separate build process. This method led to several issues:
- Concurrency problems with MSBuild, as multiple builds cannot run simultaneously.
- Overly complex synchronization requirements.

With this change we now retrieve references directly from the csproj file.
This new method is sufficient for all detectors running as part of the
MSBuild task, since none of them requires to work with transitive
dependencies.
As a result, performance is significantly improved and the process is
overall much simpler.